Update rails 3.2.11 -> 4.0.2
This is a major version update for Rails and there are many breaking
changes. The biggest one that affected this project was the removal of
attr_accessible in favor of strong parameters. This new paradigm calls
for the controller to specify which parameters are permitted from user
input.

For this change, I converted the Url model's params, since we actually
send params from the web form that we want to save on the Url model. The
Snapshot model, while it had an attr_accessible declaration did not
actually use any params from any forms, so I simply removed the
declaration. I manually ran through all of the flows and things seem to
still work.

Change-Id: I5ba9920ee48e15c3e112e761e8528fd729879e89