Introduce concept of "baseline"
After playing around with Dpxdt (another visual great diff tool) I
realized that comparing a snapshot with the previous one just isn't
enough. Dpxdt (pronounced "depicted") uses a concept of a "baseline" to
decide what snapshot to diff against.

This commit adds a button to the snapshots#show page: "Set as baseline
for this URL". On click, that snapshot is treated as the baseline for
that URL configuration. Whenever a new snapshot is generated, the
current baseline will be used to compare/diff against.

I added a basic test case for the uniqueness behavior of the Baseline
model. More tests are needed, but I intend to switch to rspec soon so
this will have to do as a first pass. I also cleaned up some fixtures in
order to get my test running.