Update attr_accessible to strong parameters
We upgraded to Rails 4 in:

  https://github.com/trotzig/diffux/commit/276f26bd3bc633efef85105f2b34e31028dfa05b

After this was merged, a bunch of commits that had been done in Rails 3
land by another author were merged. This ordering happened to cause the
app to break because the new commits made use of attr_accessible, which
was removed in Rails 4. This commit seeks to address these problems.

Change-Id: I15f0a22500d25212914ec5649daf76ac19f12ce2