Switch to Rspec for testing
I am more familiar with writing tests with Rspec than with the default
Rails implementation. @lencioni is also an Rspec guy so we're switching
over to that instead.

When I was moving over specs, I decided to ditch all the auto-generated
specs. They added little value and for the sake of getting good tests, I
decided it is better we add them manually ourselves.

I stubbed out uploading images to a CDN (Cloudinary) and the whole
Phantomjs process needed to screenshot pages. Testing that part will
have to be manual for a while. Stubbing is dealt with by mocha, also
because of familiarity.