Add Travis CI
We want to our code to be of high quality, and that means running our
automated tests with each change to make sure we didn't break anything.
Thankfully, Travis CI exists to do this work for us. All we need to do
is add a configuration file that specifies what configuration to use and
how to run the tests, and wire up Travis CI in our repo settings.

I'm starting this out by running tests on Ruby 1.9.3, 2.0.0, and 2.1.0.
We may want to adjust these in the future, but this seemed like a
reasonable set of versions to test against.

At this time, I added a badge to our readme to indicate the status of
the tests. This should help keep us accountable and also lend more
credibility to the project.

Closes #10

Change-Id: I80ad3ad036a55ef056dd0671d295b298609a3b0c