Add basic StaticPagesController spec
We want all of our pages to be assured to be in working order. This
commit adds some basic missing specs for the StaticPagesController,
which currently doesn't do much. At the same time, I added a token
`render` call to the controller action itself, to keep style with the
rest of the project.

Change-Id: I7e9c5007993056aea5b1ac2869e690a41ff0f237