Improve snapshot and image variable names
Using numbers in variable names is a code smell that can lead to
confusion down the road, especially when more meaningful names are so
easily found. Changing `snapshot1` and `image1` to `snapshot_after` and
`image_after`, and `snapshot2` and `image2` to `snapshot_before` and
`image_before`.

I had considered using `baseline` instead of `before`, but it seemed
like this class shouldn't have any knowledge of `baseline`, even if in
variable naming.

Change-Id: I5477296888f49084b8ef6f7226b16cdb57e0f2ef