Extract private method from SnapshotComparer#compare!
The compare! method was a little on the complex side, so I decided to
make it easier to parse by breaking it down a bit. This commit extracts
a private method from inside the loop that calculates the diff score
between two pixels. My intention is to make the code easier to
understand.

Change-Id: Icde585c950cfcc59182507b466a52d3f4a376037