Avoid restricting height of snapshot sample image
Restricting the height of this image caused the visual diff to be fuzzy,
making it difficult to see the parts of the page that have actually
changed. Removing this restriction may slow down the diffs, but they are
more accurate. Since we added OilyPNG recently anyway, we hope that the
slowness will be minimal, and the improvement in accuracy seems to be
well worth this tradeoff.

Change-Id: I6d552e83b3809de788b4ab57cebe6366f53c000b