Make diffs work when screenshots are different sizes
When taking snapshots of dynamic pages, or pages that have been
modified, the dimensions of the page often change. Since we are
restricting the width of our snapshots, we don't need to worry about the
width of the snapshot as much as the height.

Because we were using an algorithm that diffed the two image snapshots
based on the dimensions of the before image, this could cause problems
when the dimensions of the after image were not the same. The two common
problems noticed were:

  1) When the after image was shorter, we would run into an out of
     bounds error
  2) When the after image was longer, the difference would not appear in
     the diff

To fix this, I changed our loops to iterate over the max width and
height of the two images being diffed, and filled in the missing space
with transparent pixels. This required me to add the alpha channel to
the `pixel_diff_score` method, which was simple enough.

At the same time, I made the specs not care about the precise difference
score, since that can easily fluctuate and feels a bit like an
implementation detail. If we care to lock that in place, we can add
specific specs around that functionality at that time. However, for the
specs as written, a simple greater than zero check will be sufficient.

Change-Id: I98595b1d7f3e17d150a80aa4d744228d35bb972a