Indicate baseline snapshot in list of snapshots
As pointed out by #14, when looking at a URL that has a lot of
snapshots, it is difficult to reason about where you are at in the
review process. There should be an indication of which snapshot is the
baseline, which need to be reviewed, and which have already been
reviewed.

This commit is a quick fix for understanding where the baseline sits in
the list by giving it a green border. More work needs to be done here,
but this should resolve the major pain point.

Change-Id: I6d59743ed8c60822146a0117fa9112e6f59db54e