Pass info from PhantomJS to app when taking snapshots
We want to inspect the title of the page to determine its title, and
store that information on the snapshot soon. To prepare the app for
this, we need to be able to send information back to the app from the
take snapshot script, which runs on PhantomJS.

Since the Ruby PhantomJS object can take a block that reads lines from
STDOUT, which includes `console.log`, we can use this to pass
information from the script to the app. Since we are moving data from
JavaScript to Ruby, JSON is a great candidate.

To kick this off, I decided to include the page's title in the JSON
response in addition to the options, which had been previously just
logged to the console. I also threw in the status code for fun, since
that may end up being useful.

Change-Id: If2e915db241cceae06b48deff0d6858d230983da