Store snapshot titles
In 35e18d74a6e79d66 we added the ability for passing information from
PhantomJS to the app and included the page's title in this initial
information. We want to eventually surface the title in the app when
viewing snapshots, but before we can do that, we need to store it
somewhere. The Snapshot model seems to make perfect sense for this.

This commit adds a title property to the Snapshot model to store this
information. I also instructed the JSON parser to symbolize the names
(keys) so that we can access them via symbol instead of string. This
makes this Hash more consistent and feel more natural to work with.

Change-Id: I1f3befb7ef3bbc896e67a4e78b83724ce7bd92fb