Remove name from Url
We recently added document titles to Snapshots[1], and we plan to introduce
Projects soon[2] as a way to collect Urls. When we do this, we want the
management of Urls to be as easy as possible, perhaps via a textarea to
begin with.

This means that the Url model should be very simple as well.  Plus,
since we have the document title on Snapshots, we should be able to pull
this information into the UI if we find the need to give Urls better
names than what their addresses convey.

This commit rips out the name property from the Url model in the hope
that this will allow our future work to go more smoothly with less
effort.

[1]: https://github.com/trotzig/diffux/commit/4671bf5f1c2d581833b05098b007e192f0fc0e57
[2]: https://github.com/trotzig/diffux/issues/12#issuecomment-33337173

Change-Id: Ie39847239eba9b1fdb7df2097722aaeae13f0578