Don't upload image from SnapshotComparer
This commit frees up some of the responsibility of the SnapshotComparer
class. Instead of having it upload the diff image to Cloudinary, it now
simply returns the ChunkyPNG image.

To avoid creating and saving/uploading empty diff images I made it so
that SnapshotComparer will only return a diff image if there is a diff.
This should speed up image comparisons a little bit when no change is
detected. The Snapshot UI will look the same though.