Make RemoveNameFromUrl migration rollbackable
While working on #25, I found myself needing to rollback some migrations
that had been run. When doing so, I ran into a roadblock with this
migration because when we removed the column we did not include
information like the data type. This commit adds that info to this
migration to make it possible to rollback through this migration if we
ever need to again.

Change-Id: Icf7bb6c86b678d2cfa3865dda84167e2c9ee18a0