Strip whitespace in Project#string_to_array
When editing a Project's URLs, you could essentially duplicate a URL if
you added some whitespace around it. This is not the desired behavior,
so I modified the regex that we are using to split this input to include
any whitespace, which includes CR (\r), LF (\n), Tab (\t), and space ( )
among other things. This will help make sure that the URLs are always
actually unique and valid.

It is worth noting that the spec I added was not able to actually
reproduce the problem that I was able to cause in the browser. I'm not
sure why yet, but I figured the spec had value at least as documentation
of how we want things to work even if it doesn't seem to be properly
testing what we want.

Change-Id: Iafc9050e359cf162ae55798c95773a6ea234991f