Introduce "Pending" state for snapshots
This commit is in preparation of adding workers to do snapshotting and
snapshot comparisons. Before that can happen, we need the snapshot page
to be able to handle rendering a snapshot before it has the snapshot
image available.

I simplified the "Status?" display a bit and removed dates next to
accepted/rejected. If we need this to be around, we can add them back.
But for now, the pragmatic approach seemed to be to keep it at a simple
string.

The snapshot card will display a big timer icon in case the snapshot is
in "Pending" state.

Most of all, this commit adds specs around the different states a
snapshot can find itself in.