Use correct headings everywhere
Pages should start with h1 and work down from there in order. For now I
left all of the styling as the default Boostrap styling, but we may
decide that a lot of it is too big. Thankfully, we should be able to
easily configure this by setting variables like:

  $font-size-h2

Change-Id: Ia91c6ddd7675b02c8d7e8acc3f49c33e8b9c91a0