Hide accepted/rejected buttons when snapshot is pending
When the snapshot is pending, the image hasn't been created and diffed
against the baseline yet. So, the notion of accepting or rejecting the
diff doesn't make any sense, since the diff is not yet ready. This
commit prevents these buttons from rendering unless the snapshot is in a
state where they make sense.

Change-Id: I7034ae1bff60e0222305ef5012b05069ed0b7163