Simplify URLs that start with 'http://'
Since 'http://' is the default protocol for the Web, we can avoid
showing it in a few places. This improves the signal-to-noise ratio in
the app, and makes it more obvious when a URL is using 'https://'.

This is also consistent with how Chrome shows URLs that begin with
'http://', so people should be more or less familiar with this technique
by now.

Change-Id: I03fa5180ce57a8228c2988b9e922fd52b64019de