Shorten "Website URL" to "URL", add <abbr>
The copy "Website URL" seems a bit unnecessary to me, so I am
simplifying it to just "URL". At the same time I added an <abbr> tag
with a class of .initialism, which is a feature of bootstrap that makes
the text slightly smaller. This also improves the consistency of the
copy in the app.

Change-Id: I2cd5be8f0d72aa45ae4588444ddcc6358af33060