Allow SnapshotWorkers that can't find Snapshot to rest in peace
If you were to create a Snapshot and then delete it before the worker
had the chance to act on it, the work request would stay in the queue,
causing it to be run every time the workers spun up. This could happen
easily, for instance, if there was an unexpected error that prevented
the worker from returning successfully. In fact, this is a problem that
I saw quite often while developing Diffux.

To allow these zombie workers to rest in peace, I am adding a rescue for
the exception that is raised when the record is not found. Because, if
the Snapshot they were supposed to act upon is no longer there, there is
no more work for them to do.

Change-Id: If582e4b27514b5c7783480669db51172b4391c37