Replace Cloudinary with local solution (Paperclip)
The dependency to Cloudinary was starting to hurt us, and we didn't use
much of their features.

  - Uploading files to Cloudinary during snapshotting and comparing was
    taking a decent amount of time.
  - We had our Cloudinary key hard coded in the app instead of using ENV
    variables or some other form of config.
  - There was a risk for filling up the free Cloudinary quota.

Instead, we now rely on Paperclip (and under the hood, ImageMagick) to
do the heavy lifting of storing, resizing and serving images.

Parts of the snapshot interface changed. Instead of letting
Snapshotter#take_snapshot! send back a result I had it be in charge of
saving the image to the Snapshot model. This hurts modularity and places
more responsibility on the Snapshotter, but it was needed in order to
get a smooth workflow in regards to temp files (I don't want to leak
these files).

I've kept old Snapshot model attributes around for just a short while.
The upcoming two commits will:

  1. Introduce a manual migration tool to get all the Cloudinary images
  to Paperclip images.

  2. Get rid of the deprecated fields (along with the migration tool).

I did this so that others can sequentially update their environment
without loosing data.