Add 5 second delay in take-snapshot script
We have been seeing some problems when taking snapshots of web pages
where things happen after the page load event is fired. See #30 for more
context.

As a quick fix, we decided to add a 5 second delay after the page has
loaded before rendering the page to the image, to help allow for things
like asynchronous JavaScript to finish.

We hope to eventually implement a more robust solution that combines a
timeout with listening to request events, but this quick fix is a good
and pragmatic step in the right direction that should get us most of the
way there for now.

Change-Id: Id352b0a1f56d164b9d698a6a3d9d39dbf4bb658a