Use translucent before-image + bright color for diff
Previously we were generating a diff image that only contained grayscale
pixels that were different between the two images. While this was
effective at helping us spot that there were differences, we spent a lot
of time looking at the three images to figure out where the differences
actually were.

To help remedy this, we believe that it makes sense to render a faded
version of the "before" image (baseline) underneath the diff pixels. To
make sure that the diff pixels aren't lost on top of this translucent
image, we are rendering them in a bright pink color. We chose bright
pink because it is very noticeable and also unlikely to dominate most
snapshots.

We set the opacity of the diff pixels to be at minimum double the
opacity of the translucent image. This will help even the smallest
differences actually be noticeable.

At the same time, I changed the magic array-like access on the ChunkyPNG
images to use `get_pixel` and `set_pixel` because these methods do the
same thing as the magic array-like accessors but without checking of the
bounds. Because we are in a tight loop and because we have already
defined the bounds of the image, this should improve the performance of
the diffs without sacrificing any robustness.

Change-Id: I964ac3af1baaafbcde6d517858b9619e7b7a2723