Show diff above before/after snapshots
Now that we have a better visual diff that is overlayed on a translucent
version of the "before" image (baseline), we are actually able to
interpret the results in isolation. This means that the diff is useful
by itself. Separating these allows us to let the diff be at or much
closer to its original size, which will help when interpreting it.
Additionally, putting the before and after snapshots closer to each
other benefits a different type of visual comparison.

Eventually, we may end up baking these modes into a comparison widget,
as #9 suggests, but for now simply showing them on the same page is just
fine.

Change-Id: I023e2a659d06f40a030952ea5a036b8ed4416660