Even out before/after image widths
As you scrolled down the page, the snapshot image with the padding was
slightly vertically offset from the one without the padding, even if the
images were virtually identical. This was happening because their widths
were different by 1px, so the browser was scaling the slightly narrower
snapshot a little differently than the slightly wider one.

To fix this, I added padding to each of the elements. To allow things to
be aligned properly on both the left and right sides, I put the padding
in the middle. This doubles the gap, but since it is only 2px now,
things should be just fine.

Change-Id: I2b0575e0054c38d0bf56621c859f1f73d6e1da42