Replace "Create new snapshots" button with icon
These buttons were taking up a lot of horizontal space on the page, and
just adding significantly to visual clutter on projects that have a lot
of URLs.

In this commit, I'm trying out using an icon next to the other icons for
the URL to improve the UI of this page. While I don't really like the
whole `method` option in principle, it is convenient here.

Change-Id: Iaf9c123f7d8c76f8382e14c299551c9c4fc24379