Prevent underline on hover for URL icons
When hovering over the icons for each URL on the project page, there was
a bit of a tiny underline following the icon that looked a little out of
place. This was happening because there was whitespace inside the link,
after the icon element. To fix this, I added some HAML whitespace eaters
around these spans.

Change-Id: Iaa3d7653e22da062e2ab206cfbd3033436602613