Break up Snapshot#accept!
It felt a little weird to have two method calls in a row that both end
up saving the instance. To resolve this, I split up Snapshot#accept!
into Snapshot#accept, which simply modifies the properties, and
Snapshot#accept!, which calls #accept followed by #save!.

I had considered doing this with a conditional in the worker, but this
seemed more readable to me.

Change-Id: I604bc0c95e92eec2dc2236c38491b76bf8162852