Add heroku-multi-buildpack file for phantomjs
Since Heroku is a Read Only filesystem you need to download all the
stuff you'll need at slug compilation time. So the lazy phantomjs
downloading of the tarball won't work.

Fortunately, someone has done the work of making a package that we can
use to fetch phantomjs and include it at compile time. A subsequent
commit will configure the phantomjs gem to use it at the correct path.