Add PhantomJS shim to run on Heroku
I wish there were a way to provide a simple path to a PhantomJS
executable. But it requires a bit of implementation detail as the
PhantomJS gem selects the first usable platform.

None of the default platforms will be usable on Heroku (because of the
read-only filesystem), so this creates a new platform that points to the
correct executable, which has already been installed.