Enable Turbolinks
Turbolinks makes following links in your web application faster. Instead
of letting the browser recompile the JavaScript and CSS between each
page change, it keeps the current page instance alive and replaces only
the body and the title in the head.

This is a quick and easy drop-in speed improvement that should improve
the user experience. Since app is so simple it seems like it might be a
good fit. If we decide that we don't like it, it should be very easy to
roll back.

Change-Id: I55caafae20ab4640183ce62e7ac66016f68072da