Add loading indicator for Turbolinks
Since adding Turbolinks in adb2ca47b9f82062, I noticed that there is a
UX problem that happens whenever you click on a link that dosen't load
"instantly". This is because there is no visual indication that
something is actually happening when you click the link, so you
double-guess whether you actually clicked it properly.

To fix this, I'm adding a simple loading indicator that will appear
whenever a turbolink is clicked. We can expand this behavior in the
future if needed, but for now I wanted to keep things simple.

Change-Id: Ifc63e30aad6f998180de538b30839df348d4b8c7