Treat snapshots as pending until they have a diff
Previously, we used to treat snapshots as "under review" as soon as a
snapshot had been taken. In that, we completely ignored the fact that
most snapshots need to be diffed before they should move from "pending"
to "under review". This commit addresses that to make snapshots stay in
"pending" state a while longer.

This is a partial fix for https://github.com/trotzig/diffux/issues/46