Add Replacejs through Bower
I'm about to make the accept/reject buttons post through AJAX instead of
doing a full page load. To make this easier I'm bringing in Replacejs
[1]. And to make JavaScript package management easier I also brought in
Bower [2].

JQuery is a dependency of Replacejs. We already had JQuery in our
application manifest, and I wasn't sure what version we're on. So I
didn't use the bower-imported JQuery version that came along with
Replacejs. @lencioni might have input on this.

[1] https://github.com/causes/replacejs
[2] http://bower.io/