Add "In review" block to snapshot page
If you enter a snapshot through the "Needs review" panel on a sweep
page, you will now see a block on top of the snapshot page listing all
snapshots that need to be reviewed. I tried to keep this as unintrusive
as possible by letting the snapshot#show page take a `review_list`
parameter and only act based on that information.

The currently active snapshot will be highlighted in the panel.