Add keyboard navigation to review
This commit adds left/right key navigation to move between snapshots in
the "In review" panel on the snapshot page. We might want to switch to
"[" and "]" instead of left and right to stay more in line with e.g.
Gerrit code review. But for now left and right seem like the right
start.