Let Bower deal with jquery and jquery-ujs
Instead of using gems to pull in these JavaScript dependencies, we can
use Bower. An added bonus from this was that we upgraded jquery to
2.1.0.

I had to make some changes to the .bowerrc to get bower to work
properly. I also had to add a "resolutions" block in bower.json to
resolve what jquery version to use (jquery-ujs's bower.json specifies
jquery < 1.7, but seems to work fine with 2.1.0).