Add API endpoint to trigger sweeps
I added a similar action to the `SweepsController#create` called
`#trigger`. This action does the same thing, but sends back responses in
JSON format. If all is good, you get the URL to the new sweep back. If
there are errors, you get the errors back.

The API endpoint is meant to be triggered something like this:

    curl --header "Accept: application/json" \
         --header "Content-Type: application/json" \
         --data '{ \
                   "title": "Deploy #1", \
                   "description": "Release Notes: Fixed layout bug" \
                 }' \
         http://localhost:3000/projects/6/sweeps/trigger

This curl example block is available as copy-paste on the sweep#new page
and the sweep#index page.

Rails does some magic to be able to treat the JSON object as belonging
to the `sweep` namespace. Don't ask me how that works, it just does...

To test this, I had to write a request spec - mainly because of the JSON
parameter. More about request specs:
https://www.relishapp.com/rspec/rspec-rails/docs/request-specs/request-spec