Remove chunky_png from Gemfile
The oily_png gem, which is a Ruby C extension to speed up the pure Ruby
ChunkyPNG library, has chunky_png as a dependency. So we may as well
remove chunky_png from our Gemfile.

There is a newer version of chunky_png out (1.3.0), but oily_png hasn't
been updated yet, so we can't upgrade until that project is ready.

Change-Id: I4c0dbbf4ece857de5ccd599f0f6eeeee2b31c0f6