Center clock icons in snapshot thumbnails
The glyphicon class sets `display: inline-block;` which prevents the
element from spanning the full width of its container, which prevents
`text-align: center;` from performing its magic.

To resolve this, I added `display: block;` to this class along with
`text-align: center;`. This caused a blue underline to happen on hover,
because of how the block display mode interacts with text-decoration, so
I prevented text-decoration on hover/focus of these links.

Change-Id: I703fb5ed704d9458aacfed090c3e768ea040ec39