Make entire snapshot thumbnail box clickable
It is always best to make entire areas that look clickable actually
clickable. We were breaking this basic usability tenet with our snapshot
thumbnails, which presented a rounded rectangle with an image (that was
linked) and some caption text. Because the box make the entire thing
look clickable but it wasn't, we found ourselves constantly clicking
parts of the box that were not covered by the link.

This commit remedies this problem by making the entire box the link.

Change-Id: I28c0b718ad8281addd3b552037edca8f8e56586c