Add Diffux logo to navbar
Included in this commit is the Diffux logo, graciously constructed by
malin.trotzig@gmail.com in 20 minutes from a draft published by
@lencioni in https://github.com/trotzig/diffux/issues/31. She made the
"UX" part capital letters and used the secondary color.

The SVG provided is the raw ouput from Adobe Illustrator. It might be
possible to optimize it, I could follow up with that.