Make navbar collapse on mobile
I started out trying to figure out why the navbar was having content
that was overflowing the window width on mobile. I ended up changing the
menu to a collapsable such. On desktop, it looks the same as before. But
on smaller screens, there is a toggle button to show/hide the menu
content.

This fixes a number of issues with the menu on mobile:

 - It no longer overflows
 - The nav items now have a margin to the left edge
 - The menu was taking up too much space vertically