Disable CSS transitions, jQuery animations when snapshotting
Snapshots may end up being taken at slightly different points in the
animation, causing a visual difference when there really is no
difference. This has the potential to cause a lot of noise.

Disabling CSS transitions should prevent a lot of animations from
happening on the page.

Since many sites rely on jQuery for animations, we can reduce the amount
of noise caused by this problem by disabling jQuery animations when the
page loads, if jQuery is available. We can do this by setting $.fx.off
to true. More info:

  http://api.jquery.com/jQuery.fx.off/

There may be other things that we can do to reduce the noise caused by
animations, but this seems like a pretty good start considering how many
sites rely on a combination of CSS and jQuery to do animations.

Change-Id: Ic9e3b9285b67fd1309984ee2d1ac4aea3d947e8e