Fix accepted snapshot showing as pending
To reproduce the bug:

  1. Create a new URL in a project and trigger a first snapshot for that
  URL
  2. Accept the snapshot when it's ready
  3. Create a second snapshot for the same URL
  4. Accept it too.
  5. Go back to the first snapshot. It will show up as "Pending".

This was caused by the (overly complicated? I'm to blame...) logic for
when a snapshot is considered to be `waiting_for_diff?`. If there was a
baseline out there, and no diff image had been produced, we would treat
the snapshot as pending. I fixed that by adding a check to make sure
that the baseline is older than the current snapshot. It might not work
in all possible cases, but I think it is good enough for now.