Change page:before-change event to page:fetch
The page:before-change event is apparently only fired from Turbolinks
when a link is actually clicked. Since we are accessing the Turbolinks
API directly, when handling keyboard shortcuts, we are loading new pages
but were not seeing the loading indicator. To fix this, I am changing
this event listener from the page:before-change event to the next one in
line, page:fetch. This event is triggered when Turbolinks starts to
fetch the target page, according to their documentation.

Change-Id: I9e36355e22a8f14fe60094f9ecc33687c2afbb40