Replace "view all" link next to viewport
When viewing a project that has some viewports and shapshots, we were
rendering a link that read "View all". To reduce clutter, I decided to
simply link the viewport text itself.

One drawback of this method is Bootstrap will render this text identical
to non-links in this element except for the :hover state (which gets
underlined), which may make it more difficult to discover that these are
links. However, since it is most likely that most people will only care
about the latest couple of snapshots anyway, this probably isn't that
big of a deal. The benefit of simplifying the visually complex interface
a little bit seems to outweigh this possible drawback.

Change-Id: Ic4c6814b3f61d56a81bd81d23aad5bad68316cd4